services/auth-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
python-multipart==0.0.6
pydantic[email]==2.5.2
